Merge
보근님 구현 기능과
제작한 구현 기능 병함
(실 데이터 연동)

+++ Assets/02.Scripts/Managers/ResourceManager.cs
@@ -41,7 +41,7 @@ public class ResourceManager : MonoBehaviour
 
     public void UpdateTotalLifeIncreaseUI()
     {
-        int totalLifeIncrease = 0;
+        float totalLifeIncrease = 0;
         foreach (var root in roots)
         {
             totalLifeIncrease += root.baseLifeGeneration;
